Tests + fix

+++ src/main/java/telegram/teamjob/service/PetPhotoServiceImpl.java
@@ -28,7 +28,10 @@ public class PetPhotoServiceImpl implements PetPhotoService {
 
     @Value("${path.to.photos.folder}")
     private String photoDis;
+    @Value("${telegram.bot.token}")
+    private String value;
     private final PetPhotoRepository petPhotoRepository;
+
     private final RecordRepository recordRepository;
 
     public PetPhotoServiceImpl(PetPhotoRepository petPhotoRepository, RecordRepository recordRepository) {
@@ -39,39 +42,44 @@ public class PetPhotoServiceImpl implements PetPhotoService {
     public void uploadPhoto(Long recordId, PhotoSize[] photos) throws IOException {
         logger.info("Was invoked method for upload avatar");
         Record record = recordRepository.findByRecordId(recordId);
-        String fileId = photos[0].fileId();
-        String fileName = photos[0].fileUniqueId();
-        System.out.println(fileId);
-        System.out.println(fileName);
+        if (record != null) {
+            String fileId = photos[0].fileId();
+            String fileName = photos[0].fileUniqueId();
+            System.out.println(fileId);
+            System.out.println(fileName);
 
-        String botToken = "5607238907:AAGr_a_h_GqQKemI2z_JSKPxsSfwiqbvemM";
-        URL url = new URL("https://api.telegram.org/bot" + botToken + "/" + "getFile?file_id=" + fileId);
-        BufferedReader br = new BufferedReader(new InputStreamReader(url.openStream()));
-        String getFileResponse = br.readLine();
+            String botToken = value;
+            URL url = new URL("https://api.telegram.org/bot" + botToken + "/" + "getFile?file_id=" + fileId);
+            BufferedReader br = new BufferedReader(new InputStreamReader(url.openStream()));
+            String getFileResponse = br.readLine();
 
-        JsonElement jResult = JsonParser.parseString(getFileResponse);
-        JsonObject path = jResult.getAsJsonObject().getAsJsonObject("result");
-        String file_path = path.get("file_path").getAsString();
-        System.out.println(file_path);
+            JsonElement jResult = JsonParser.parseString(getFileResponse);
+            JsonObject path = jResult.getAsJsonObject().getAsJsonObject("result");
+            String file_path = path.get("file_path").getAsString();
+            System.out.println(file_path);
 
-        Path filePath = Path.of(photoDis, record.getRecordId() + "." +
-                getExtensions(Objects.requireNonNull(fileName)));
-        Files.createDirectories(filePath.getParent());
-        Files.deleteIfExists(filePath);
-        try (
-                InputStream is = new URL("https://api.telegram.org/file/bot" + botToken + "/" + file_path).openConnection().getInputStream();
-                OutputStream os = Files.newOutputStream(filePath, CREATE_NEW);
-                BufferedInputStream bis = new BufferedInputStream(is, 480);
-                BufferedOutputStream bos = new BufferedOutputStream(os, 480);
-        ) {
-            bis.transferTo(bos);
+            Path filePath = Path.of(photoDis, record.getRecordId() + "." +
+                    getExtensions(Objects.requireNonNull(fileName)));
+            Files.createDirectories(filePath.getParent());
+            Files.deleteIfExists(filePath);
+            try (
+                    InputStream is = new URL("https://api.telegram.org/file/bot" + botToken + "/" + file_path).openConnection().getInputStream();
+                    OutputStream os = Files.newOutputStream(filePath, CREATE_NEW);
+                    BufferedInputStream bis = new BufferedInputStream(is, 480);
+                    BufferedOutputStream bos = new BufferedOutputStream(os, 480);
+            ) {
+                bis.transferTo(bos);
+            }
+            PetPhoto petPhoto = new PetPhoto();
+            petPhoto.setRecord(record);
+            petPhoto.setFilePath(filePath.toString());
+            petPhoto.setFileSize(photos[0].fileSize());
+            br.close();
+            petPhotoRepository.save(petPhoto);
+            logger.info("Photo was saved");
+        } else {
+            logger.info("Record not found");
         }
-        PetPhoto petPhoto = new PetPhoto();
-        petPhoto.setRecord(record);
-        petPhoto.setFilePath(filePath.toString());
-        petPhoto.setFileSize(photos[0].fileSize());
-        br.close();
-        petPhotoRepository.save(petPhoto);
     }
     @Override
     public String getExtensions(String fileName) {
